Split existing tests into unit and integration, add tags and add to CI
