+++ ably/http_paginated_response_integration_test.go
@@ -1,3 +1,6 @@
+//go:build !unit
+// +build !unit
+
 package ably_test
 
 import (
